Small additions

+++ src/llm/Embedding.cpp
@@ -1,4 +1,4 @@
-#include "llm/embedding.h"
+#include "llm/Embedding.h"
 
 Embedding::Embedding(const int vocabSize, const int dim, const int contextLength) : tokenEmbedding(vocabSize, dim), gradientToken(vocabSize, dim), positionEmbedding(contextLength, dim), gradientPosition(contextLength, dim), dim(dim), contextLength(contextLength) {
 	tokenEmbedding.randomize();
@@ -12,7 +12,7 @@ Tensor Embedding::forward(const std::vector<int>& inputTokens) {
 	const size_t n = inputTokens.size();
 	const Tensor tok(n, dim);
 	const Tensor pos(n, dim);
-	for (int i = 0; i < n; i++) {
+	for (size_t i = 0; i < n; i++) {
 		tok.setRow(i, tokenEmbedding.getRow(inputTokens[i]));
 		pos.setRow(i, positionEmbedding.getRow(i));
 	}
